Added login functionality to backend

+++ part4/bloglist/controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -18,12 +20,16 @@ blogRouter.get('/:id', async (request, response) => {
 
 blogRouter.post('/', async (request, response) => {
   const body = request.body
-
+  const token = request.body.authorization
   if (body.title === undefined || body.url === undefined){
     response.status(400).end()
   }
   else{
-    const user = await User.findById(body.userId)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
     const blog = new Blog({
       'url': body.url,
       'title': body.title,
@@ -49,7 +55,18 @@ blogRouter.put('/:id', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const token = request.body.token
+  const decodedtoken = jwt.verify(request.token, process.env.SECRET)
+  if (!token || !decodedtoken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const delBlog = await Blog.findById(request.params.id)
+
+  if (delBlog.author.toString() === decodedtoken.id.toString() ){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }else{
+    response.status(400).end()
+  }
 })
 module.exports = blogRouter
